+++ src/shims.ts
@@ -1,3 +1,1 @@
-// @ts-ignore
 export { fetch, Response, Request, Headers } from '@sveltejs/kit/install-fetch';
-