+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 - initial draft
 
+## 0.1.1
+
+- updated to the latest sveltekit 1.0.0-next.225
+
+- improve the render response to include status/headers
+
+- fixed build hang issue
+
 ### Patch Changes
 
 - various